increase grow/shrink distance limits
--automatic icon sizer is using a prediction
algorithm in order to decide to grow or not the
items size. There is a chance that the prediction
calculations are executed too early and as
such grow/shrink limits it is better to not
be to near to each other. That approach should
make the AutomaticItemSizer implementation a little
more robust.