fix dock view geometry calculations
--when background custom radius and shadow
are applied the dock window geometry shouldnot
change if these values are not altered. This
is especially true when a dock is hiding
its screen edge margin when there is a maximized
window and the same time is hiding also its
background shadows. In such case in the past because
shadows were temporarily disabled the dock was constantly
changing its geometry