do not start sliding ins/outs with no reason
--ViewPart:VisibilityManager is sending events
to show or hide but it can not validate the view
visual state meaning is hidden, sliding in/out, blocked
etc. That role is taken from Containment::VisibilityManager
that accepts that signals and in case the state are
requesting is already provided then in such case
these signals are ignored.
--This way myView flags for inSlidingIn/Out/PartiallyShown/FullyShown
etc are providing valid values in some previous
problematic scenarios
--This fix solves the issue that parabolic effect with Tasks Previews
it was not too responsive but it had many glitches